+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 
 ## 0.3.3
 
-- Update glob to regex conversion to prevent command name matching directory components 
+- `init --with-examples` now copies all new bundled example files.
+- Update glob to regex conversion to prevent command name matching directory components.
 - Handle missing rich dependency, prevent potential ImportErrors when rendering output.
-- Fix error in help for logs subcommand
-- Fix `ValueError` on Ctrl-C when exiting `logs --follow` mode
+- Fix error in help for logs subcommand.
+- Fix `ValueError` on Ctrl-C when exiting `logs --follow` mode.
 
 ## 0.3.2